Documentado HamburguesaService y HamburguesaDTO

+++ src/main/java/org/spdgroup/bigbitebackend/model/entities/DetalleInsumo.java
@@ -19,5 +19,5 @@ public class DetalleInsumo {
     @JoinColumn(name = "insumo_id", nullable = false)
     private Insumo insumo;
 
-    private int cantidad;
+    private double cantidad;
 }
